example/requirements.txt
aiosqlite
asyncpg
fastapi[all]
greenlet
httpx
pytest_asyncio
sqlalchemy
uvicorn